Added both updates versions of the python code
Full U-Net code in Python (CA-UNet) includes the option (when uncommenting the right part) to choose whether to use Coordinate Attentive blocks. 
Full U-Net code in Python (Finding the only the missing band) includes a typo in the title, but is also the code that has takes the sparse frequency as an input and tries to predict just the missing part of the frequency spectra.