Uploading U-Net Python scripts
Uploading two similar Python scripts that implement the same U-Net, but filter the frequency content at different stages. Perhaps in the future it would be smart to simplify them into the same script by using a function. This way the user can select when they wish to filter out the frequencies.