Python version sparse solver
Two versions of the sparse solver approach. The estimates the frequency content of a single trace. The second makes an estimate for multiple traces, however it does not work properly and needs to be fixed. The Full U-Net code that is being uploaded is an implementation where multiple neighbouring channels are given to the neural network as additional information. The downside to this method is that each training sample's size is three times larger, which limits the total number of training data that can be given to the network (due to the hardware bottlenecking).